pyproject.toml
[project]
name = "channel-weaver"
version = "0.1.0"
description = "Python CLI tool to process multitrack WAV recordings from mixing consoles. De-interleaves channels, concatenates splits, creates stereo buses, and outputs numbered DAW-ready files. Perfect for live sessions from consoles like Midas/Behringer."
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.3.5",
    "pydantic>=2.12.4",
    "pysoundfile>=0.9.0.post1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "tqdm>=4.67.1",
    "typer>=0.20.0",
]

[dependency-groups]
dev = [
    "hypothesis>=6.148.7",
    "mypy>=1.19.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-github-actions-annotate-failures>=0.3.0",
    "pytest-mock>=3.15.1",
    "pytest-xdist>=3.8.0",
    "ruff>=0.14.9",
    "types-pyyaml>=6.0.12.20250915",
    "types-tqdm>=4.67.0.20250809",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (may use real filesystem)",
    "slow: Tests that take more than 1 second",
    "ffmpeg: Tests that involve FFmpeg mocking",
    "pydantic: Tests for Pydantic model validation",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__pycache__/*", "**/conftest.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.mypy]
ignore_missing_imports = true
